tests: try to remove the DB again if it failed
This has been observed on the EL5 build machine quite a few times.

+++ tests/sql/util-delete-database.sh
@@ -4,6 +4,12 @@ cd `dirname $0`
 
 . ./util-config.sh
 
-psql -q -U $DESKA_SU -c "DROP DATABASE IF EXISTS ${DESKA_DB};" || die "Drop database"
+# Dropping the DB occasionally fails on EL5, so let's try it multiple times
+psql -q -U $DESKA_SU -c "DROP DATABASE IF EXISTS ${DESKA_DB};"
+if [[ $? -ne 0 ]]; then
+    sleep 5
+    psql -q -U $DESKA_SU -c "DROP DATABASE IF EXISTS ${DESKA_DB};" || die "Drop database"
+fi
+
 psql -q -U $DESKA_SU -c "DROP USER IF EXISTS ${DESKA_USER};" || die "Drop user"
 psql -q -U $DESKA_SU -c "DROP USER IF EXISTS ${DESKA_ADMIN};" || die "Drop user"
